refactor: style improvements and code refactor

+++ src/app/components/single-user/single-user.component.ts
@@ -1,7 +1,6 @@
-import { Component, Input, OnDestroy, OnInit } from '@angular/core';
-import { FormControl, FormGroup } from '@angular/forms';
-import { ActivatedRoute, ParamMap, Router } from '@angular/router';
-import { Subscription, switchMap } from 'rxjs';
+import { Component, OnDestroy, OnInit } from '@angular/core';
+import { ActivatedRoute, Router } from '@angular/router';
+import { Subscription } from 'rxjs';
 import { Permission, User } from 'src/app/models';
 import { ApiService } from 'src/app/services/api.service';
 import { MyAuthService } from 'src/app/services/my-auth.service';
@@ -14,14 +13,12 @@ import { MyAuthService } from 'src/app/services/my-auth.service';
 export class SingleUserComponent implements OnInit, OnDestroy {
   private subscription!: Subscription;
   user!: User;
+
+  firstname: string = '';
+  lastname: string = '';
+  email: string = '';
+  password: string = '';
   permissionList: Permission[] = [];
-  
-  editUserForm = new FormGroup({
-    firstname: new FormControl(''),
-    lastname: new FormControl(''),
-    email: new FormControl(''),
-    password: new FormControl('')
-  });
   
   authorisedToDelete: boolean = false;
   authorisedToUpdate: boolean = false;
@@ -34,37 +31,60 @@ export class SingleUserComponent implements OnInit, OnDestroy {
     this.subscription = this.route.params.subscribe(params => {
       this.apiService.findUserByEmail(params['email']).subscribe((user) =>{
         this.user = user;
+        this.firstname = user.firstname;
+        this.lastname = user.lastname;
+        this.email = user.email;
+        this.password = user.password;
         this.permissionList = user.permissionList;
       })
     });
   }
-
-  ngOnDestroy(): void {
-    this.subscription.unsubscribe();
-  }
-
-  editUser() {
-    const firstname = this.editUserForm.get('firstname')?.value;
-    const lastname = this.editUserForm.get('lastname')?.value;
-    const email = this.editUserForm.get('email')?.value;
-    const password = this.editUserForm.get('password')?.value;
-
-    this.apiService.updateUser(firstname!, lastname!, email!, password!, this.permissionList)
+  
+  editUser(formData: any) {
+    this.firstname = formData.firstname;
+    this.lastname = formData.lastname;
+    this.email = formData.email;
+    this.password = formData.password;
+    this.permissionList = formData.permissionList;
+    
+    if (this.isValidData()) {
+      this.apiService.updateUser(this.firstname, this.lastname, this.email, this.password, this.permissionList)
       .subscribe((updatedUser) => {
         this.user = updatedUser;
       });
+    }
   }
-
-  refreshPermissionList(permissions: Permission[]) {
-    this.permissionList = permissions;
+  
+  isValidData() {
+    if (
+      this.firstname !== null &&
+      this.firstname !== '' &&
+      this.lastname !== null &&
+      this.lastname !== '' &&
+      this.email !== null &&
+      this.email !== '' &&
+      this.password !== null &&
+      this.password !== ''
+    ) {
+      return true;
+    }
+    return false;
   }
-
+  
   deleteUser(){ 
+    if(this.user.email === localStorage.getItem('signedInUserEmail')) {
+      alert('You can not delete yourself.');
+      return;
+    }
     if(confirm('Are you sure you want to delete this user?')){
       this.apiService.deleteUser(this.user.email)
       .subscribe(() =>  {
         this.router.navigate(['/users']);
       })
     }
+  }
+
+  ngOnDestroy(): void {
+    this.subscription.unsubscribe();
   }
 }
